chore: unified extensible structure
