+++ vllm_omni/diffusion/models/wan2_2/pipeline_wan2_2.py
@@ -12,7 +12,6 @@ from typing import Any, cast
 
 import PIL.Image
 import torch
-from diffusers import FlowMatchEulerDiscreteScheduler
 from diffusers.utils.torch_utils import randn_tensor
 from torch import nn
 from transformers import AutoTokenizer, UMT5EncoderModel
@@ -23,6 +22,7 @@ from vllm_omni.diffusion.distributed.autoencoders.autoencoder_kl_wan import Dist
 from vllm_omni.diffusion.distributed.cfg_parallel import CFGParallelMixin
 from vllm_omni.diffusion.distributed.utils import get_local_device
 from vllm_omni.diffusion.model_loader.diffusers_loader import DiffusersPipelineLoader
+from vllm_omni.diffusion.models.dmd2 import DMD2PipelineMixin
 from vllm_omni.diffusion.models.progress_bar import ProgressBarMixin, _is_rank_zero
 from vllm_omni.diffusion.models.schedulers import FlowUniPCMultistepScheduler
 from vllm_omni.diffusion.models.wan2_2.scheduling_wan_euler import WanEulerScheduler
@@ -955,95 +955,6 @@ def _load_model_index(model: str, local_files_only: bool) -> dict:
         except Exception:
             pass
     return {}
-
-
-class DMD2EulerScheduler(FlowMatchEulerDiscreteScheduler):
-    """Euler scheduler that always uses the fixed DMD2 training timestep schedule."""
-
-    def __init__(self, *args, dmd2_timesteps: list[int], **kwargs):
-        super().__init__(*args, **kwargs)
-        self._dmd2_timesteps = dmd2_timesteps
-
-    def set_timesteps(
-        self,
-        num_inference_steps: int | None = None,
-        device: str | torch.device | None = None,
-        **kwargs,
-    ) -> None:
-        super().set_timesteps(timesteps=self._dmd2_timesteps, device=device)
-
-
-class DMD2PipelineMixin:
-    """Mixin for FastGen DMD2-distilled models. Must appear before the base pipeline in MRO."""
-
-    def __init_dmd2__(self) -> None:
-        """Call after super().__init__() to apply DMD2 scheduler and read model_index."""
-        local_files_only = os.path.exists(self.od_config.model)
-        model_index = _load_model_index(self.od_config.model, local_files_only)
-
-        dmd2_timesteps = model_index.get("dmd2_denoising_timesteps", [999, 937, 833, 624])
-        self.num_inference_steps = model_index.get("dmd2_num_inference_steps", 4)
-        shift = model_index.get("dmd2_scheduler_shift", 1.0)
-        self.dmd2_guidance_scale = model_index.get("dmd2_guidance_scale", 1.0)
-
-        self.scheduler = DMD2EulerScheduler(
-            num_train_timesteps=1000,
-            shift=shift,
-            dmd2_timesteps=dmd2_timesteps,
-        )
-
-    def _sanitize_dmd2_request(self, req: OmniDiffusionRequest) -> None:
-        """Sanitize CFG-related fields in-place. Mutates req.sampling_params and req.prompts."""
-        sp = req.sampling_params
-
-        if sp.num_inference_steps and sp.num_inference_steps != self.num_inference_steps:
-            logger.warning(
-                "DMD2: ignoring num_inference_steps=%d, forcing %d.",
-                sp.num_inference_steps,
-                self.num_inference_steps,
-            )
-        sp.num_inference_steps = self.num_inference_steps
-
-        if sp.guidance_scale_provided and sp.guidance_scale != self.dmd2_guidance_scale:
-            logger.warning(
-                "DMD2: ignoring guidance_scale=%.2f, forcing %.2f.",
-                sp.guidance_scale,
-                self.dmd2_guidance_scale,
-            )
-        sp.guidance_scale = self.dmd2_guidance_scale
-        sp.guidance_scale_provided = False
-
-        if sp.guidance_scale_2 is not None:
-            logger.warning("DMD2: ignoring guidance_scale_2.")
-            sp.guidance_scale_2 = None
-
-        if sp.true_cfg_scale is not None:
-            logger.warning("DMD2: ignoring true_cfg_scale.")
-            sp.true_cfg_scale = None
-
-        sp.do_classifier_free_guidance = False
-        sp.is_cfg_negative = False
-
-        fixed = []
-        for p in req.prompts:
-            if isinstance(p, dict) and "negative_prompt" in p:
-                logger.warning("DMD2: ignoring negative_prompt.")
-                p = {k: v for k, v in p.items() if k != "negative_prompt"}
-            fixed.append(p)
-        req.prompts = fixed
-
-    def forward(self, req: OmniDiffusionRequest, **kwargs) -> DiffusionOutput:
-        self._sanitize_dmd2_request(req)
-        # Safety: remove DMD2-controlled params from kwargs to avoid TypeError
-        # if a caller passes them explicitly alongside **kwargs.
-        kwargs.pop("guidance_scale", None)
-        kwargs.pop("num_inference_steps", None)
-        return super().forward(
-            req,
-            guidance_scale=self.dmd2_guidance_scale,
-            num_inference_steps=self.num_inference_steps,
-            **kwargs,
-        )
 
 
 class WanT2VDMD2Pipeline(DMD2PipelineMixin, Wan22Pipeline):
